moved browser plugin reports to an own plugin 'DevicePlugins'
