+++ core/Updates/2.10.0-b10.php
@@ -1,0 +1,49 @@
+<?php
+/**
+ * Piwik - free/libre analytics platform
+ *
+ * @link http://piwik.org
+ * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ *
+ */
+
+namespace Piwik\Updates;
+
+use Piwik\DataAccess\ArchiveTableCreator;
+use Piwik\Updater;
+use Piwik\Updates;
+
+class Updates_2_10_0_b10 extends Updates
+{
+
+    static function getSql()
+    {
+        $sqls = array();
+
+        $archiveTables = ArchiveTableCreator::getTablesArchivesInstalled();
+
+        $archiveBlobTables = array_filter($archiveTables, function($name) {
+            return ArchiveTableCreator::getTypeFromTableName($name) == ArchiveTableCreator::BLOB_TABLE;
+        });
+
+        foreach ($archiveBlobTables as $table) {
+
+            $sqls["UPDATE " . $table . " SET name = 'DevicePlugins_plugin' WHERE name = 'UserSettings_plugin'"] = false;
+        }
+
+        return $sqls;
+    }
+
+    static function update()
+    {
+        $pluginManager = \Piwik\Plugin\Manager::getInstance();
+
+        try {
+            $pluginManager->activatePlugin('DevicePlugins');
+        } catch(\Exception $e) {
+        }
+
+        Updater::updateDatabase(__FILE__, self::getSql());
+    }
+
+}
