+++ plugins/DevicePlugins/Columns/PluginJava.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  *
  */
-namespace Piwik\Plugins\UserSettings\Columns;
+namespace Piwik\Plugins\DevicePlugins\Columns;
 
 use Piwik\Common;
 use Piwik\Plugin\Dimension\VisitDimension;
@@ -14,9 +14,9 @@ use Piwik\Tracker\Request;
 use Piwik\Tracker\Visitor;
 use Piwik\Tracker\Action;
 
-class PluginQuickTime extends VisitDimension
+class PluginJava extends VisitDimension
 {
-    protected $columnName = 'config_quicktime';
+    protected $columnName = 'config_java';
     protected $columnType = 'TINYINT(1) NOT NULL';
 
     /**
@@ -27,6 +27,6 @@ class PluginQuickTime extends VisitDimension
      */
     public function onNewVisit(Request $request, Visitor $visitor, $action)
     {
-        return Common::getRequestVar('qt', 0, 'int', $request->getParams());
+        return Common::getRequestVar('java', 0, 'int', $request->getParams());
     }
 }
