+++ plugins/DevicePlugins/Reports/GetPlugin.php
@@ -6,11 +6,11 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  *
  */
-namespace Piwik\Plugins\UserSettings\Reports;
+namespace Piwik\Plugins\DevicePlugins\Reports;
 
 use Piwik\Piwik;
 use Piwik\Plugin\ViewDataTable;
-use Piwik\Plugins\UserSettings\Columns\Plugin;
+use Piwik\Plugins\DevicePlugins\Columns\Plugin;
 
 class GetPlugin extends Base
 {
@@ -18,18 +18,18 @@ class GetPlugin extends Base
     {
         parent::init();
         $this->dimension     = new Plugin();
-        $this->name          = Piwik::translate('UserSettings_WidgetPlugins');
-        $this->documentation = Piwik::translate('UserSettings_WidgetPluginsDocumentation', '<br />');
+        $this->name          = Piwik::translate('DevicePlugins_WidgetPlugins');
+        $this->documentation = Piwik::translate('DevicePlugins_WidgetPluginsDocumentation', '<br />');
         $this->metrics       = array('nb_visits');
         $this->constantRowsCount = true;
         $this->processedMetrics = array('nb_visits_percentage');
         $this->order = 4;
-        $this->widgetTitle  = 'UserSettings_WidgetPlugins';
+        $this->widgetTitle  = 'DevicePlugins_WidgetPlugins';
     }
 
     public function configureView(ViewDataTable $view)
     {
-        $this->getBasicUserSettingsDisplayProperties($view);
+        $this->getBasicDevicePluginsDisplayProperties($view);
 
         $view->config->addTranslations(array(
             'label'                => $this->dimension->getName(),
@@ -43,7 +43,7 @@ class GetPlugin extends Base
         $view->config->show_all_views_icons    = false;
         $view->config->show_table_all_columns  = false;
         $view->config->columns_to_display  = array('label', 'nb_visits_percentage', 'nb_visits');
-        $view->config->show_footer_message = Piwik::translate('UserSettings_PluginDetectionDoesNotWorkInIE');
+        $view->config->show_footer_message = Piwik::translate('DevicePlugins_PluginDetectionDoesNotWorkInIE');
 
         $view->requestConfig->filter_sort_column = 'nb_visits_percentage';
         $view->requestConfig->filter_sort_order  = 'desc';
