+++ plugins/DevicePlugins/Visitor.php
@@ -6,9 +6,9 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  *
  */
-namespace Piwik\Plugins\UserSettings;
+namespace Piwik\Plugins\DevicePlugins;
 
-require_once PIWIK_INCLUDE_PATH . '/plugins/UserSettings/functions.php';
+require_once PIWIK_INCLUDE_PATH . '/plugins/DevicePlugins/functions.php';
 
 class Visitor
 {
