+++ plugins/DevicePlugins/lang/en.json
@@ -1,0 +1,10 @@
+{
+    "DevicePlugins": {
+        "BrowserWithNoPluginsEnabled": "%1$s with no plugins enabled",
+        "BrowserWithPluginsEnabled": "%1$s with plugins %2$s enabled",
+        "PluginDescription": "Reports browser plugins",
+        "PluginDetectionDoesNotWorkInIE": "Note: Plugins detection doesn't work in Internet Explorer. This report is only based on non-IE browsers.",
+        "WidgetPlugins": "Browser Plugins",
+        "WidgetPluginsDocumentation": "This report shows which browser plugins your visitors had enabled. This information might be important for choosing the right way to deliver your content."
+    }
+}
