+++ plugins/UserSettings/Archiver.php
@@ -25,7 +25,6 @@ require_once PIWIK_INCLUDE_PATH . '/plugins/UserSettings/functions.php';
 class Archiver extends \Piwik\Plugin\Archiver
 {
     const LANGUAGE_RECORD_NAME = 'UserSettings_language';
-    const PLUGIN_RECORD_NAME = 'UserSettings_plugin';
 
     const LANGUAGE_DIMENSION = "log_visit.location_browser_lang";
 
@@ -35,7 +34,6 @@ class Archiver extends \Piwik\Plugin\Archiver
      */
     public function aggregateDayReport()
     {
-        $this->aggregateByPlugin();
         $this->aggregateByLanguage();
     }
 
@@ -45,32 +43,9 @@ class Archiver extends \Piwik\Plugin\Archiver
     public function aggregateMultipleReports()
     {
         $dataTableRecords = array(
-            self::PLUGIN_RECORD_NAME,
             self::LANGUAGE_RECORD_NAME,
         );
         $this->getProcessor()->aggregateDataTableRecords($dataTableRecords, $this->maximumRows);
-    }
-
-    protected function aggregateByPlugin()
-    {
-        $selects = array(
-            "sum(case log_visit.config_pdf when 1 then 1 else 0 end) as pdf",
-            "sum(case log_visit.config_flash when 1 then 1 else 0 end) as flash",
-            "sum(case log_visit.config_java when 1 then 1 else 0 end) as java",
-            "sum(case log_visit.config_director when 1 then 1 else 0 end) as director",
-            "sum(case log_visit.config_quicktime when 1 then 1 else 0 end) as quicktime",
-            "sum(case log_visit.config_realplayer when 1 then 1 else 0 end) as realplayer",
-            "sum(case log_visit.config_windowsmedia when 1 then 1 else 0 end) as windowsmedia",
-            "sum(case log_visit.config_gears when 1 then 1 else 0 end) as gears",
-            "sum(case log_visit.config_silverlight when 1 then 1 else 0 end) as silverlight",
-            "sum(case log_visit.config_cookie when 1 then 1 else 0 end) as cookie"
-        );
-
-        $query = $this->getLogAggregator()->queryVisitsByDimension(array(), false, $selects, $metrics = array());
-        $data = $query->fetch();
-        $cleanRow = LogAggregator::makeArrayOneColumn($data, Metrics::INDEX_NB_VISITS);
-        $table = DataTable::makeFromIndexedArray($cleanRow);
-        $this->insertTable(self::PLUGIN_RECORD_NAME, $table);
     }
 
     protected function aggregateByLanguage()
