+++ plugins/UserSettings/Controller.php
@@ -10,7 +10,7 @@ namespace Piwik\Plugins\UserSettings;
 
 use Piwik\Plugins\Resolution\Reports\GetConfiguration;
 use Piwik\Plugins\UserSettings\Reports\GetLanguage;
-use Piwik\Plugins\UserSettings\Reports\GetPlugin;
+use Piwik\Plugins\DevicePlugins\Reports\GetPlugin;
 use Piwik\Plugins\Resolution\Reports\GetResolution;
 use Piwik\View;
