+++ plugins/UserSettings/lang/en.json
@@ -1,8 +1,6 @@
 {
     "UserSettings": {
         "BrowserLanguage": "Browser language",
-        "BrowserWithNoPluginsEnabled": "%1$s with no plugins enabled",
-        "BrowserWithPluginsEnabled": "%1$s with plugins %2$s enabled",
         "Language_aa": "Afar",
         "Language_ab": "Abkhazian",
         "Language_ae": "Avestan",
@@ -189,9 +187,6 @@
         "Language_zu": "Zulu",
         "LanguageCode": "Language code",
         "PluginDescription": "Reports various User Settings: Browser, Browser Family, Operating System, Plugins, Resolution, Global Settings.",
-        "PluginDetectionDoesNotWorkInIE": "Note: Plugins detection doesn't work in Internet Explorer. This report is only based on non-IE browsers.",
-        "VisitorSettings": "Visitor Settings",
-        "WidgetPlugins": "Browser Plugins",
-        "WidgetPluginsDocumentation": "This report shows which browser plugins your visitors had enabled. This information might be important for choosing the right way to deliver your content."
+        "VisitorSettings": "Visitor Settings"
     }
 }
